Force blocks to be downloaded in assending order
Using the `sparse` option shouldn't be necessary, but if not used, the
processing will slow way down until all the blocks have been downloaded.
This is because the feed doesn't prioritize missing blocks based on
where the readable feed is reading.

There's an issue to fix this in hypercore already:
https://github.com/mafintosh/hypercore/issues/76